Script extended to auto generate bug reports
The script now searches for existing bug reports and creates new ones if needed in to the TF_Bug_Report_Holding repo to be sorted into actual bugs and test code refactors.

If a report is closed, it will be reopened and new fail data added to the comments.

If a report is open and at least 7 days old it will then create  anew comment on the bug report stating the issue is still present in the latest runtime, also listing the runtime version.