workflow validation update
This update enables only runs from Beta, Monthly and Red to use the TF Compare script. There is no real need for it to be run every time a dev pushes up a small fix as it is not guaranteed there would be a run exactly the same to compare against, unlike 2 Beta RT runs.