Update: Logic added to download the last 10 runs to then compare
The idea is that currently only downloading the last 2 runs means that the script will break if the last 2 runs failed for some reason.

Now, we download the last 10 runs, in which we pull out the one that is currently running and a previous run that has a status of success. This will ensure a consistent ability to compare runs.