Code refactor to ensure artifacts are downloaded correctly and the script continues to run
The script should never fall over, the only time it should exit and not complete is if an artifact in the current run hasn't been created, however, if that is the case, the TF compare script will never start to run in the first place because the workflow would have already failed before reaching the TF compare script.

If in the last 2 runs (current run included) have 2 valid artifact files, they will be downloaded and compared, if the previous run doesn't have an artifact file because maybe the run crashed then the script will continue to download the artifact file from the current run and report the fails as bug reports as usual.