Remove get workflow run logs and add RTVersion
New Runtime version variable has been added to GitHub, this is now passed in to the job command line as a parameter which is now accessed within the python script.

The old way of getting the RT version via downloading the logs has now been removed, making the script more efficient and accurate.